requirements.txt
google-cloud-pubsub
google-cloud-storage
autorootcwd
telethon
rich
PyYaml
pytz
pytest
pytest-asyncio